requirements.txt
colorama
python-dotenv
aiogram
asyncpg
#pymysql
pyyaml
